gnocchi-indexer-sqlalchemy not packaged
gnocchi-indexer-sqlalchemy is not longer packaged for Ubuntu
which is the reason scenario001 is failing CI right now.

This patch removes the installation of the gnocchi-indexer-sqlalchemy
package on Debian based platforms but still leaves it intact for Red Hat
based.

I have reached out to James and Corey from Canonical for confirmation if
this is permanent or not ported, should not be merged until confirmed.

Change-Id: I74d93e1d7012c43431259bd8335df17650b8242f